only swagger left

+++ controllers/learningcenter.controller.js
@@ -5,20 +5,42 @@ import {
 } from "../validations/learningcenter.validation.js";
 import Region from "../models/region.model.js";
 import User from "../models/user.model.js";
-import { Op } from "sequelize";
+import { Op, fn, col, literal } from "sequelize";
 import LCField from "../models/lcfields.model.js";
 import Field from "../models/field.model.js";
 import { loggerError, loggerInfo } from "../logs/logger.js";
 import Like from "../models/like.model.js";
 import Comment from "../models/comment.model.js";
+import Branch from "../models/branch.model.js";
+
+const deleteOldImage = (imgPath) => {
+  if (imgPath) {
+    const fullPath = path.join("uploads", imgPath);
+    if (fs.existsSync(fullPath)) {
+      fs.unlinkSync(fullPath);
+    }
+  }
+};
 
 async function findAll(req, res) {
   try {
     let all = await LearningCenter.findAll({
+      attributes: {
+        include: [
+          [
+            literal(`(
+              SELECT COUNT(*) 
+              FROM Likes 
+              WHERE Likes.learningCenterId = LearningCenter.id
+            )`),
+            "numberOfLikes",
+          ],
+        ],
+      },
       include: [
         {
           model: Branch,
-          attributes: [ "id", "name", "img", "regionId", "phoneNumber", "address",],
+          attributes: ["id", "name", "img", "regionId", "phoneNumber", "address"],
         },
         {
           model: Region,
@@ -33,21 +55,19 @@ async function findAll(req, res) {
           attributes: ["id", "name", "professionId", "subjectId"],
         },
         {
-          model: Like,
-          attributes: ["id", "userId", "learningCenterId"],
-        },
-        {
           model: Comment,
           attributes: ["id", "comment", "userId", "star"],
         },
       ],
+      subQuery: false, 
     });
+
 
     if (all.length === 0) {
       loggerError.error(
         `ERROR: No information available.;  Method: ${req.method};  LearningCenter-FindAll`
       );
-      return res.status(404).json({ error: "No information available." });
+      return res.status(404).json({ error: "Learning Centers Not Found" });
     }
 
     loggerInfo.info(
@@ -57,22 +77,34 @@ async function findAll(req, res) {
     res.status(200).send({ data: all });
   } catch (e) {
     loggerError.error(
-      `ERROR: ${e};  Method: ${req.method};  LearningCenters-FindAll`
+      `ERROR: ${e.message};  Method: ${req.method};  LearningCenters-FindAll`
     );
 
-    res.status(500).json({ error: e });
+    res.status(500).json({ error: e.message });
   }
 }
-
 async function findOne(req, res) {
   try {
     let { id } = req.params;
+
     let one = await LearningCenter.findOne({
       where: { id },
+      attributes: {
+        include: [
+          [
+            literal(`(
+              SELECT COUNT(*) 
+              FROM Likes 
+              WHERE Likes.learningCenterId = LearningCenter.id
+            )`),
+            "numberOfLikes",
+          ],
+        ],
+      },
       include: [
         {
           model: Branch,
-          attributes: [ "id", "name", "img", "regionId", "phoneNumber", "address",],
+          attributes: ["id", "name", "img", "regionId", "phoneNumber", "address"],
         },
         {
           model: Region,
@@ -87,40 +119,37 @@ async function findOne(req, res) {
           attributes: ["id", "name", "professionId", "subjectId"],
         },
         {
-          model: Like,
-          attributes: ["id", "userId", "learningCenterId"],
-        },
-        {
           model: Comment,
           attributes: ["id", "comment", "userId", "star"],
         },
       ],
+      subQuery: false, 
     });
 
     if (!one) {
       loggerError.error(
-        `ERROR: Region Not Found;  Method: ${req.method};  LearningCenter-FindOne`
+        `ERROR: Learning Center Not Found;  Method: ${req.method};  LearningCenter-FindOne`
       );
-      return res.status(404).json({ error: "Region Not Found" });
+      return res.status(404).json({ error: "Learning Center Not Found" });
     }
 
     loggerInfo.info(
-      `Method: ${req.method};  Saccessfully FindOne LearningCenter;`
+      `Method: ${req.method};  Successfully FindOne LearningCenter;`
     );
 
     res.status(200).json({ data: one });
   } catch (e) {
     loggerError.error(
-      `ERROR: ${e};  Method: ${req.method};  LearningCenters-FindOne`
+      `ERROR: ${e.message};  Method: ${req.method};  LearningCenters-FindOne`
     );
 
-    res.status(500).json({ error: e });
+    res.status(500).json({ error: e.message });
   }
 }
 
 async function create(req, res) {
   try {
-    req.body.userId = req.user.id;
+    req.body.createdBy = req.user.id;
     let { error, value } = LearningCenterValidation.validate(req.body);
 
     if (error) {
@@ -140,13 +169,24 @@ async function create(req, res) {
       );
       return res
         .status(409)
-        .json({ error: "There is a learning center with this name." });
+        .json({ error: "Learning Center with this name already exists!" });
+    }
+
+    let checkRegion = await Region.findByPk(value.regionId);
+
+    if (!checkRegion) {
+      loggerError.error(
+        `There is no such region with this ID;  Method: ${req.method};  LearningCenter-Create`
+      );
+      return res
+        .status(404)
+        .json({ error: "There is no such region with this ID" });
     }
 
     let LcField = value.lcfield || [];
 
-    for (let e of LcField) {
-      let check = await Field.findOne({ where: { id: e } });
+    for (let field of LcField) {
+      let check = await Field.findOne({ where: { id: field } });
 
       if (!check) {
         loggerError.error(
@@ -154,23 +194,31 @@ async function create(req, res) {
         );
         return res
           .status(404)
-          .json({ error: "There are no such IDs in the Field table!" });
+          .json({ error: "There is no such field available with this ID" });
       }
     }
 
     let newLc = await LearningCenter.create(value);
 
-    LcField.forEach(async (e) => {
-      await LCField.create({ fieldId: e, learningCenterId: newLc.id });
+    if (!newLc) {
+      loggerError.error(
+        `ERROR: Learning Center Not Created;  Method: ${req.method};  LearningCenter-Create`
+      );
+      return res.status(400).json({ error: "Learning Center Not Created.Please try again" });
+    }
+
+    LcField.forEach(async (field) => {
+      await LCField.create({ fieldId: field, learningCenterId: newLc.id });
     });
 
     loggerInfo.info(
       `Method: ${req.method};  Saccessfully Create LearningCenter;`
     );
-    res.status(201).json({ message: "Created Succesfully" });
+
+    res.status(201).json({ message: "New Learning Center Created Succesfully" });
   } catch (e) {
     loggerError.error(
-      `ERROR: ${e};  Method: ${req.method};  LearningCenters-Create`
+      `ERROR: ${e.message};  Method: ${req.method};  LearningCenters-Create`
     );
 
     res.status(500).json({ error: e.message });
@@ -179,6 +227,7 @@ async function create(req, res) {
 
 async function update(req, res) {
   try {
+
     let { id } = req.params;
 
     let check = await LearningCenter.findOne({ where: { id } });
@@ -189,6 +238,16 @@ async function update(req, res) {
       );
       return res.status(404).json({ error: "Learning Center Not Found" });
     }
+
+    if (req.user.role != "ADMIN" && check.createdBy != req.user.id) {
+      loggerError.error(
+        `ERROR: You are not allowed to update this Learning Center;  Method: ${req.method};  LearningCenter-Update`
+      );
+      return res
+        .status(403)
+        .json({ error: "You are not allowed to update this Learning Center" });
+    }
+
     let { error, value } = LearningCenterPatchValidation.validate(req.body);
 
     if (error) {
@@ -198,17 +257,34 @@ async function update(req, res) {
       return res.status(400).json({ error: error.details[0].message });
     }
 
+    if (value.regionId) {
+      let checkRegion = await Region.findByPk(value.regionId);
+
+      if (!checkRegion) {
+        loggerError.error(
+          `There is no such region with this ID;  Method: ${req.method};  LearningCenter-Update`
+        );
+        return res
+          .status(404)
+          .json({ error: "There is no such region with this ID" });
+      }
+    }
+
+    if (value.img) {
+      deleteOldImage(check.img);
+    }
+
     await LearningCenter.update(value, { where: { id } });
     loggerInfo.info(
       `Method: ${req.method};  Saccessfully Update LearningCenter;`
     );
-    res.status(200).json({ message: "Update Successfully" });
+    res.status(200).json({ message: "Learning Center Updated Successfully" });
   } catch (e) {
     loggerError.error(
-      `ERROR: ${e};  Method: ${req.method};  LearningCenters-Update`
+      `ERROR: ${e.message};  Method: ${req.method};  LearningCenters-Update`
     );
 
-    res.status(500).json({ error: e });
+    res.status(500).json({ error: e.message });
   }
 }
 
@@ -221,24 +297,38 @@ async function remove(req, res) {
       loggerError.error(
         `ERROR: Learning Center Not Found;  Method: ${req.method};  LearningCenter-Delete`
       );
-      return res.status(401).json({ error: "Learning Center Not Found" });
+      return res.status(404).json({ error: "Learning Center Not Found" });
+    }
+
+    if (req.user.role != "ADMIN" && check.createdBy != req.user.id) {
+      loggerError.error(
+        `ERROR: You are not allowed to delete this Learning Center;  Method: ${req.method};  LearningCenter-Update`
+      );
+      return res
+        .status(403)
+        .json({ error: "You are not allowed to delete this Learning Center" });
+    }
+
+    if (value.img) {
+      deleteOldImage(check.img);
     }
 
     await LearningCenter.destroy({ where: { id } });
     loggerInfo.info(
       `Method: ${req.method};  Saccessfully Delete LearningCenter;`
     );
-    res.status(201).json({ message: "Delete Successfully" });
+    res.status(201).json({ message: "Learning Center Deleted Successfully" });
   } catch (e) {
     loggerError.error(
-      `ERROR: ${e};  Method: ${req.method};  LearningCenters-Delete`
+      `ERROR: ${e.message};  Method: ${req.method};  LearningCenters-Delete`
     );
-    res.status(401).json({ error: e });
+    res.status(500).json({ error: e.message });
   }
 }
 
 async function Search(req, res) {
   try {
+    
     let { page, take, regionId } = req.query;
 
     if (page || take) {
@@ -249,7 +339,49 @@ async function Search(req, res) {
       let categories = await LearningCenter.findAndCountAll({
         limit: take,
         offset: offset,
+        attributes: {
+          include: [
+            [
+              literal(`(
+                SELECT COUNT(*) 
+                FROM Likes 
+                WHERE Likes.learningCenterId = LearningCenter.id
+              )`),
+              "numberOfLikes",
+            ],
+          ],
+        },
+        include: [
+          {
+            model: Branch,
+            attributes: ["id", "name", "img", "regionId", "phoneNumber", "address"],
+          },
+          {
+            model: Region,
+            attributes: ["id", "name"],
+          },
+          {
+            model: User,
+            attributes: ["id", "firstName", "lastName", "role"],
+          },
+          {
+            model: Field,
+            attributes: ["id", "name", "professionId", "subjectId"],
+          },
+          {
+            model: Comment,
+            attributes: ["id", "comment", "userId", "star"],
+          },
+        ],
+        subQuery: false, 
       });
+
+      if (categories.rows.length === 0) {
+        loggerError.error(
+          `ERROR: Learning Center Not Found;  Method: ${req.method};  LearningCenter-Search`
+        );
+        return res.status(404).json({ error: "Learning Centers Not Found" });
+      }
 
       return res.status(200).json({
         totalItems: categories.count,
@@ -271,7 +403,7 @@ async function Search(req, res) {
       }
     });
 
-    console.log(regionId);
+  
     if (regionId) {
       let regionIds = Array.isArray(regionId) ? regionId : [regionId];
       conditions.regionId = {
@@ -289,21 +421,26 @@ async function Search(req, res) {
     let results = await LearningCenter.findAll({
       where: conditions,
       order: order.length > 0 ? order : [["id", "ASC"]],
+      attributes: {
+        include: [
+          [
+            literal(`(
+              SELECT COUNT(*) 
+              FROM Likes 
+              WHERE Likes.learningCenterId = LearningCenter.id
+            )`),
+            "numberOfLikes",
+          ],
+        ],
+      },
       include: [
+        {
+          model: Branch,
+          attributes: ["id", "name", "img", "regionId", "phoneNumber", "address"],
+        },
         {
           model: Region,
           attributes: ["id", "name"],
-        },
-        {
-          model: Branch,
-          attributes: [
-            "id",
-            "name",
-            "img",
-            "regionId",
-            "phoneNumber",
-            "address",
-          ],
         },
         {
           model: User,
@@ -314,15 +451,19 @@ async function Search(req, res) {
           attributes: ["id", "name", "professionId", "subjectId"],
         },
         {
-          model: Like,
-          attributes: ["id", "userId", "learningCenterId"],
-        },
-        {
           model: Comment,
           attributes: ["id", "comment", "userId", "star"],
         },
       ],
+      subQuery: false, 
     });
+
+    if (results.length === 0) {
+      loggerError.error(
+        `ERROR: Learning Center Not Found;  Method: ${req.method};  LearningCenter-Search`
+      );
+      return res.status(404).json({ error: "Learning Centers Not Found" });
+    }
 
     loggerInfo.info(
       `Method: ${req.method};  Successfully Search LearningCenter; data: ${results}`
